Refactor, UIViewControllerTransitioningDelegate객체를 따로 생성하지 않고 ViewController가 해당 책임 수행하도록 변경
