+++ Projects/Feature/DesignSystem/Feature/Sources/Components/Alert/AlertPresentable.swift
@@ -9,8 +9,6 @@ import UIKit
 
 public protocol AlertPresentable: AnyObject {
     
-    var alertTransitionDelegate: UIViewControllerTransitioningDelegate? { get set }
-    
     func presentAlert(presentingController: UIViewController, listener: DSTwoButtonAlertViewControllerListener, config: DSTwoButtonAlert.Config)
     
     func dismissAlert(presentingController viewController: UIViewController, completion: (()->Void)?)
