+++ Projects/Feature/DesignSystem/Feature/Sources/Components/Alert/Type2/DSTwoButtonAlertViewController.swift
@@ -12,7 +12,7 @@ public protocol DSTwoButtonAlertViewControllerListener: AnyObject {
     func action(_ action: DSTwoButtonAlertViewController.Action)
 }
 
-public final class DSTwoButtonAlertViewController: UIViewController {
+public final class DSTwoButtonAlertViewController: UIViewController, UIViewControllerTransitioningDelegate {
     
     // Action
     public enum Action {
@@ -61,6 +61,59 @@ public final class DSTwoButtonAlertViewController: UIViewController {
         alertView.snp.makeConstraints { make in
             make.centerY.equalToSuperview()
             make.horizontalEdges.equalTo(view.safeAreaLayoutGuide).inset(33.5)
+        }
+    }
+}
+
+// MARK: UIViewControllerTransitioningDelegate
+public extension DSTwoButtonAlertViewController {
+    
+    func animationController(forDismissed dismissed: UIViewController) -> (any UIViewControllerAnimatedTransitioning)? {
+        CustomAlertDismissAnimation()
+    }
+    
+    func animationController(forPresented presented: UIViewController, presenting: UIViewController, source: UIViewController) -> (any UIViewControllerAnimatedTransitioning)? {
+        CustomAlertPresentAnimation()
+    }
+}
+    
+
+// MARK: CustomAlertPresentAnimation
+fileprivate final class CustomAlertPresentAnimation: NSObject, UIViewControllerAnimatedTransitioning {
+    func transitionDuration(using transitionContext: UIViewControllerContextTransitioning?) -> TimeInterval {
+        return 0.2
+    }
+    
+    func animateTransition(using transitionContext: UIViewControllerContextTransitioning) {
+        guard let toView = transitionContext.view(forKey: .to) else { return }
+        let containerView = transitionContext.containerView
+        
+        toView.alpha = 0
+        containerView.addSubview(toView)
+        
+        UIView.animate(withDuration: transitionDuration(using: transitionContext), animations: {
+            toView.alpha = 1
+        }) { _ in
+            transitionContext.completeTransition(!transitionContext.transitionWasCancelled)
+        }
+    }
+}
+
+
+// MARK: CustomAlertDismissAnimation
+fileprivate final class CustomAlertDismissAnimation: NSObject, UIViewControllerAnimatedTransitioning {
+    func transitionDuration(using transitionContext: UIViewControllerContextTransitioning?) -> TimeInterval {
+        return 0.2
+    }
+    
+    func animateTransition(using transitionContext: UIViewControllerContextTransitioning) {
+        guard let fromView = transitionContext.view(forKey: .from) else { return }
+        
+        UIView.animate(withDuration: transitionDuration(using: transitionContext), animations: {
+            fromView.alpha = 0
+        }) { _ in
+            fromView.removeFromSuperview()
+            transitionContext.completeTransition(!transitionContext.transitionWasCancelled)
         }
     }
 }
